Split into lib and cli crates
Fixes https://github.com/sunng87/datafusion-postgres/issues/21
