+++ Cargo.toml
@@ -1,19 +1,11 @@
-[package]
-name = "datafusion-postgres"
-version = "0.1.0"
-edition = "2021"
-rust-version = "1.73"
+[workspace]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+members = ["datafusion-postgres", "datafusion-postgres-cli"]
 
-[dependencies]
+[workspace.dependencies]
 pgwire = "0.23"
-datafusion = { version = "39", features = ["avro"] }
-futures = "0.3"
-async-trait = "0.1"
-tokio = { version = "1", features = ["full"] }
-chrono = { version = "0.4", features = ["std"] }
-structopt = { version = "0.3", default-features = false }
+datafusion = { version = "39", default-features = false }
+tokio = { version = "1", default-features = false }
 
 [profile.release]
 strip = true
