+++ datafusion-postgres/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "datafusion-postgres"
+version = "0.1.0"
+edition = "2021"
+rust-version = "1.73"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+pgwire = { workspace = true }
+datafusion = { workspace = true }
+futures = "0.3"
+async-trait = "0.1"
+tokio = { workspace = true, features = ["sync"] }
+chrono = { version = "0.4", features = ["std"] }
