+++ datafusion-postgres/src/lib.rs
@@ -1,0 +1,4 @@
+mod datatypes;
+mod handlers;
+
+pub use handlers::{DfSessionService, HandlerFactory, Parser};
